+++ src/callbacks/allowlists/MerkleAllowlist.sol
@@ -104,7 +104,7 @@ contract MerkleAllowlist is BaseCallback {
         bytes calldata callbackData_
     ) internal virtual {}
 
-    function _onClaimProceeds(uint96, uint96, uint96, bytes calldata) internal pure override {
+    function _onClaimProceeds(uint96, uint256, uint96, bytes calldata) internal pure override {
         // Not implemented
         revert Callback_NotImplemented();
     }
