+++ src/callbacks/liquidity/DTL.sol
@@ -62,7 +62,7 @@ abstract contract DirectToLiquidity is BaseCallback {
 
     function _onClaimProceeds(
         uint96 lotId_,
-        uint96 proceeds_,
+        uint256 proceeds_,
         uint96 refund_,
         bytes calldata callbackData_
     ) internal virtual override {
