+++ src/interfaces/ICallback.sol
@@ -79,7 +79,7 @@ interface ICallback {
     /// @dev If the Callback is configured to receive tokens, then the proceeds and/or refund will be sent prior to the call.
     function onClaimProceeds(
         uint96 lotId,
-        uint96 proceeds,
+        uint256 proceeds,
         uint96 refund,
         bytes calldata callbackData
     ) external returns (bytes4);
